Teensy code for base
This uses Sabertooth 2x12 (larger units are exactly the same to program) and "simplified serial" mode to control motors. I couldn't get the angular boost to be sufficient with ROS and diff drive control, so this takes the cmd_vel message and boosts it and sends it to motors. This robot will turn like she means it! It also publishes encoder messages and IMU, MPU6050.